:zap: improve: roles actions

+++ src/actions/roles/remove-permission.ts
@@ -3,6 +3,7 @@
 import { revalidatePath } from "next/cache"
 
 import prismadb from "@/lib/prismadb"
+import { has } from "@/lib/rbac"
 
 interface IDeleteRolePermission {
   roleId: number
@@ -14,14 +15,21 @@ export async function removePermission({
   permissionId,
 }: IDeleteRolePermission) {
   try {
+    const isAuthorized = await has({ role: "admin" })
+
+    if (!isAuthorized) {
+      return { success: false, message: "Unauthorized" }
+    }
+
     await prismadb.rolePermission.deleteMany({
       where: { roleId, permissionId },
     })
 
-    revalidatePath(`/admin/roles/permissions`)
+    revalidatePath(`/admin/roles/${roleId}/permissions`)
 
     return { success: true }
   } catch (error) {
-    return { success: false }
+    console.error("Error removing permission:", error)
+    return { success: false, message: "Something went wrong" }
   }
 }
